flask/requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
psycopg2-binary==2.8.6
requests==2.22.0